Split script into a few Python modules
